Rename left arm packages for clarity
