+++ src/dmbot_serial_left_arm/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.0.2)
-project(dmbot_serial_leftarm)
+project(dmbot_serial_left_arm)
 
 
 ## Compile as C++11, supported in ROS Kinetic and newer
