+++ src/legged_examples/left_arm_dm_hw/launch/neck_dm_hw.launch
@@ -3,13 +3,18 @@
     <arg name="ns" default="" doc="Namespace for the hardware node"/>
 
     <group ns="$(arg ns)">
-        <param name="legged_robot_description" command="$(find xacro)/xacro $(find wanren_arm)/urdf/wanren_arm.urdf"
-            robot_type:=$(arg robot_type) "/>
+        <param name="legged_robot_description"
+               command="$(find xacro)/xacro $(find wanren_arm)/urdf/wanren_arm.urdf robot_type:=$(arg robot_type)"/>
         <param name="manual_topic_override" value="true"/>
         <param name="manual_cmd_timeout" value="0.2"/>
-        <rosparam file="$(find leftarm_dm_hw)/config/$(arg robot_type).yaml" command="load"/>
+        <rosparam file="$(find left_arm_dm_hw)/config/$(arg robot_type).yaml" command="load"/>
 
-        <node name="leftarm_dm_hw" pkg="leftarm_dm_hw" type="leftarm_dm_hw" respawn="false" clear_params="true" output="screen">
+        <node name="left_arm_dm_hw"
+              pkg="left_arm_dm_hw"
+              type="left_arm_dm_hw"
+              respawn="false"
+              clear_params="true"
+              output="screen">
         </node>
     </group>
