+++ src/legged_examples/left_arm_dm_hw/package.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 <package format="2">
-    <name>leftarm_dm_hw</name>
+    <name>left_arm_dm_hw</name>
     <version>0.0.0</version>
     <description>ROS control warped interface for Unitree Robots</description>
     <maintainer email="contact@mdmbot.com">dm</maintainer>
